feat: Add dashboard menu items and header icon change
Added new menu items for "Dashboard", "Analytics", "Reports" and "Mail" inside the Drawer component. Updated the Drawer component with AccountCircle icon instead of Divider and created a commented section for future use. Replaced MenuIcon with FilterDramaIcon inside the Header component. Changed the color to primary for the IconButton. Updated headerButton in the colorConfig file to have a white font color.

refactor: Reorganize elements in the Login and Home page

Wrapped the Navigate component inside a div in both the Login and Home page. Imported react-toastify library and included a comment about possible error message display.